+++ src/main/java/dev/rynwllngtn/agorasystem/entities/account/accounts/AccountChecking.java
@@ -1,6 +1,6 @@
 package dev.rynwllngtn.agorasystem.entities.account.accounts;
 
-import dev.rynwllngtn.agorasystem.test.AccountRequest;
+import dev.rynwllngtn.agorasystem.dtos.AccountRequest;
 import dev.rynwllngtn.agorasystem.entities.account.Account;
 import dev.rynwllngtn.agorasystem.entities.user.User;
 import dev.rynwllngtn.agorasystem.enums.account.AccountType;